changed pool hash rate to float and add commit/rollback to pow insert fn

+++ postgres/models/pow_data_test.go
@@ -427,7 +427,7 @@ func testPowDataSelect(t *testing.T) {
 }
 
 var (
-	powDatumDBTypes = map[string]string{`Time`: `integer`, `NetworkHashrate`: `integer`, `PoolHashrate`: `bigint`, `Workers`: `integer`, `NetworkDifficulty`: `double precision`, `CoinPrice`: `character varying`, `BTCPrice`: `character varying`, `Source`: `character varying`}
+	powDatumDBTypes = map[string]string{`Time`: `integer`, `NetworkHashrate`: `integer`, `PoolHashrate`: `double precision`, `Workers`: `integer`, `NetworkDifficulty`: `double precision`, `CoinPrice`: `character varying`, `BTCPrice`: `character varying`, `Source`: `character varying`}
 	_               = bytes.MinRead
 )
